mindsdb hacktoberfest/use-cases/auto-banking-customer-service/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
coverage==7.10.7
decorator==5.2.1
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.115.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
jira==3.6.0
jiter==0.11.0
mindsdb_sdk==3.4.8
mindsdb_sql_parser==0.0.2
numpy==2.3.3
openai==2.3.0
packaging==25.0
pandas==2.3.3
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
sseclient-py==1.8.0
starlette==0.41.3
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.0
validators==0.20.0
